refactor: MatchResult의 static 메서드 분리
