+++ src/main/java/view/InputUntilValid.java
@@ -9,15 +9,12 @@ public interface InputUntilValid<T> {
     T validate(Supplier<T> supplier);
 
     static <T> T validatePlayerAnswer(final Player player, final Function<Player, T> function) {
-        {
-            while (true) {
-                try {
-                    return function.apply(player);
-                } catch (IllegalArgumentException e) {
-                    System.out.println(e.getMessage());
-                }
+        while (true) {
+            try {
+                return function.apply(player);
+            } catch (IllegalArgumentException e) {
+                System.out.println(e.getMessage());
             }
         }
-
     }
 }
